+++ JunctionPointer/Controllers/MainController.cs
@@ -5,6 +5,7 @@ using DirLinker.Interfaces.Controllers;
 using DirLinker.Interfaces;
 using DirLinker.Data;
 using System.Diagnostics;
+using System.ComponentModel;
 
 namespace DirLinker.Controllers
 {
@@ -29,7 +30,15 @@ namespace DirLinker.Controllers
 
         public Form Start()
         {
-            Process.Start("ElevatedWorker.exe");
+            try
+            {
+                Process.Start("ElevatedWorker.exe");
+            }
+            catch (Win32Exception E)
+            {
+                MessageBox.Show("Could not start elevated worker.\r\n" + E.Message, "DirLinker");
+                Environment.Exit(0);
+            }
 
             _operationData = new LinkOperationData();
